requirements.txt
numpy
matplotlib
pandas
scikit-learn
opencv-python
pillow
captum
ttach
scipy